requirements.txt
# 核心依赖
yt-dlp>=2023.12.30
httpx>=0.25.0
aiofiles>=23.0.0
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 翻译 (可选)
deep-translator>=1.11.0

# 任务调度
apscheduler>=3.10.0
celery>=5.3.0

# Web UI
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# 工具
loguru>=0.7.0
tenacity>=8.2.0
click>=8.1.0
rich>=13.0.0

# ============================================
# 国内大模型API SDK (按需安装)
# ============================================

# 阿里云 - DashScope (通义千问 + 语音)
dashscope>=1.14.0

# 阿里云 - 语音服务
alibabacloud-nls-java-sdk>=2.0.0

# 智谱AI
zhipuai>=2.0.0

# 讯飞语音
# xfyun-python-sdk>=1.0.0  # 需从官网下载

# ============================================
# 语音识别/TTS SDK
# ============================================

# OpenAI (Whisper API + GPT)
openai>=1.0.0

# Azure TTS
azure-cognitiveservices-speech>=1.34.0

# 阿里云 NLS
# aliyun-python-sdk-core>=2.13.0
# aliyun-python-sdk-nls>=1.0.0

# ============================================
# 视频处理 (本地FFmpeg必需)
# ============================================

moviepy>=1.0.3

# 可选: 浏览器自动化 (如需模拟发布)
# playwright>=1.40.0
